refactor(directory)
- 관심사 별로 디렉토리 세분화
