+++ src/components/TodoTemplate.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import TodoHeader from './TodoHeader';
-import TodoList from './TodoList';
+import TodoHeader from './TodoHeader/TodoHeader';
+import TodoList from './TodoBody/TodoList';
 
 export default function TodoTemplate({ todoList, onClickAdd, onClickRemove }) {
   return (
